Integrate fs metadata
